[Ninja] Parse, warn, and ignore -l and -d options.
The -l and -d options are supported by Ninja and used by scripts that
call into Ninja, but are not yet supported by llbuild's ninja driver.

For now, document them as being not implemented, parse them, warn that
they are not being used, then ignore them. This allows using scripts
that expect all Ninja options to be available immediately with llbuild.