Stale file removal fixes for Swift 4.0 (#176)
* Use `set` instead of `unordered_set`

Turns out that `std::set_difference` does not work with unordered sets,
but this didn't become apparent with the smaller sets we are using in
the unit tests.

* Fix `BuildSystemTaskTests`

This one needed to be updated after changes to
`MockBuildSystemDelegate`.

* Allow stale file removal build values (#167)

* Add delegate support for command reporting (#158)

Commands can now report additional errors, notes and warnings, not tied
to a command process as before. This is used by the `StaleFileRemoval`
command to report deleted files and demotes all its errors to warnings.

* Extract the `isLocatedUnderRootPath` check (#169)

Previously, we were incrementing the pointer with the first `res.second++`, so the next part of the if statement would see the character after that.

This extracts that code, fixes the issue and adds unit tests.