[BuildSystem] Produced nodes should build when prior value was missing
In situations where a node was previously a missing input, that fact may
be recorded in the build database. If subsequent builds add a rule that
may now produce that node, the system could get into a situation such
that 'missing input' errors are triggered even though we can
successfully build it. This changes produced nodes to cause their rule
to be built.

rdar://problem/40376404