[Tracing] Introduce os_signpost/os_log
This commit introduces the usage of os_signpost and os_log to implement tracing in llbuild. os_signpost is part of macOS 10.14 and gives the ability to add more and structured information to tracing intervals.
We adopt this in llbuild to get a better understanding about the build system if tracing as a feature has been enabled. If tracing is disabled this should result in no-ops to not influence the performance of the system.

rdar://problem/40310911