[BuildSystem] Add discovered-dependencies-file support to CAPIExternalCommand
Adds support for specifying a dependency-info-style discovered-dependencies file in a `deps` attribute on external C API commands.  This allows external commands to emit discovered-dependency information in the same way that shell commands can.

This also adds the first unit test for the BuildSystem C API.

In a future enhancement it would be useful to extend the C API to let the external command add discovered dependencies directly, without having to go through a separate file.