Fix use-after-free in some helper methods
llvm::SmallString's str() method returns an llvm::StringRef, which does
not copy the string data. When getShortDescription/getVerboseDescription
return, the SmallString is deallocated and the underlying memory freed,
causing the returned StringRef to be pointing to deallocated memory.

Remove these functions and require the caller to provide their own
string data instance, which the Command can then populate with data.