[BuildSystem] Fix tests and comments due to #328 and #329
Reverting directory contents behavior broke tests expecting the new
behavior. This change adds a test that directly exposes the bug that was
fixed in #328, and fixes the tests that were XFAIL'd in #329. It also
adds commentary explaining the problem in BuildSystem.cpp, as the
full solution is non-trivial.

rdar://problem/41182818