[Core] Set a busy timeout on the database
This adds a timeout for sqlite to wait before giving up because the
database is already busy. This happens if you have 2 concurrent builds
happening at once. Previously it would error out immediately, now it
will wait for up to 20 seconds before erroring out. This behavior is
similar to that of cargo's.