[BuildSystem] Allow processes to release lanes
The general approach here is that LaneBasedExecutionQueue::executeProcess()
exposes a pipe file descriptor to the launched process. If that process
sends the value of LLBUILD_TASK_ID over that pipe, the process will
release the lane on which it is executing and asynchronously wait for
completion.

This initial implementation spawns a new thread for each process that
triggers the release. Longer term, more sophisticated handling of async
IO and process wait may be preferred.

Since any process may opt in to use this functionality, the chain of
methods that lead into the executeProcess call have all been modified
to take completion/result reporting callbacks.

rdar://problem/44683736