[libllbuildSwift] Remove version protection from Swift bindings
The versioning is not really very useful as there is not really
a revlock issue. The Swift bindings are built together with the
C bindings and we don't have any intention of using a different version
of libllbuild with libllbuildSwift. This commit hardcodes the library
version in llbuild.h and removes the generation scripts and targets.