[BuildSystem] Split out filtered directory nodes
This change moves filtered directory support into a parallel tree of
nodes. It also introduces a new 'stat' node that produces the stat
information for a file system object.  The combination of these changes
allows filtered directories to work as intended, depending upon the
filter contents only, and not the stat info of the containing directory.

Supports rdar://problem/30640904