Use Clang attributes to import llbuild enumerations more nicely in Swift
This obviates the need to duplicate every llbuild enum in the Swift
bindings, where we may forget to update the mapping. It also provides
the Swift compiler with a real enumeration instead of a series of
integers, which allows it to emit diagnostics in cases where a switch
does not cover all enum cases, for example.

rdar://problem/46476230