[libllbuild] Added a build engine delegate callback for reporting cycles.
This callback provides all of the keys involved in the cycle, starting with the key that was initially requested up to the first repeated key.
This is to allow clients of the build engine component to report cycles as errors when they occur instead of crashing.

This is part of <rdar://problem/43870963>.