Fix ArchiveShellCommand's verbose description output.
`ArchiveShellCommand::getVerboseDescription()` had wasn't correctly using the `bool first` variable it created.
It was set to `true`,  but the loop only ever checked if it was `false` and, if so, it would insert a space and
set it to false. Since it never was false, the verbose description would never contain spaces between arguments.

This change makes this code do what I imagine was the intent: insert a space _before_ all arguments except the
first.