Added basic Swift interface for build database
The build database is not accessible from the Swift interface. While that's okay for building we might want to make it accessible to make calculations on the information we have in the database.
Added a non-destructive version of the sqlite database which doesn't re-create the database automatically if versions mismatch. In this case, an error is thrown instead.

rdar://50298712