build: pull in `FindSQLite3.cmake` fron CMake 3.14
CMake 3.14 now distributes `FindSQLite3.cmake` as a module to enable
`find_package(SQLite3)`.  This uses the standard CMake system for
managing the dependency   This allows for implicit handling for SQLite3
when installed on the system and allows for the default CMake style
parameters for overriding the search paths and link paths.