[BuildEngine] Fix potential deadlock during cancel
If a failure occurs updating the build DB during finished task
processing, we trigger synchronous cancellation of the build. This
currently triggers a deadlock, waiting for the task that failed to be
marked as finished. This change prevents that deadlock by decrementing
the outstanding task count prior to such cancellation.

rdar://problem/50203529