libllbuild: make build succeed on Windows x86_64
Windows does not support VLAs.  The current pattern would fail to build
with a failure to construct the local array as the size was not
constant.  Switch to a reserved vector which avoids the growth but
ensures that we can allocate a variable sized array.  This repairs the
Windows build.