libllbuild: fix definition of `LLBUILD_EXPORT`
The Windows build requires more careful handling of the exports macro.
WHen building the library as a static library, no dll storage should be
specified.  When building the library dynamic, DLL storage must be
applied.  This corrects the exported interfaces from the executable
(`llbuild`) exporting the interfaces for `libllbuild`.  Unix would
prevent the exporting of the symbols as `-rdynamic` is not the default.