Basic: correct dependencies for linking on Windows
This corrects the link dependencies for Windows.  llbuildBasic is always
built as a static library.  This requires that the dependencies be
linked PUBLIC so that users of the library are able to have the
dependencies fulfilled.  llbuild now properly links in ShLwApi (the
dependency was previously silently fulfilled by the implicitly linked
libraries from CMake).