Shim llvm::sys::getDefaultTargetTriple() to Swift
The default target triple for the host machine can be useful information when constructing command line arguments for a tool, but this information isn’t in LLVM’s C API, so it’s inaccessible from Swift. This commit exposes it in libllbuild as `llb_get_default_target_triple()` and in llbuildSwift as `BuildSystem.defaultTargetTriple`.