SR-11730 Improved GraphViz output
Refactored the GraphViz output to build up a set of edges frist, with the style options to highlight the critical path, and label the verticies with something more meaningful.

Changed the default behaviour to only include the critical path, and added a commandline option to include the all build keys.