Change TaskInterface to be passed by value
Clients of the core must use the TaskInterface to communicate with the
core from other threads. Passing by reference to a stack allocated
instance forces clients to be aware of this and take care copy it. This
is awkward and error prone, and also unnecessary for an object
containing two pointers.