llbuildSwift: correct a memory corruption issue
The memory allocation for the data here is done via copying the data
from the bytes in the string.  This allocation is done by
`UnsafeMutableBufferPointer.allocate`.  Map the release for that data to
`UnsafeMutableBufferPointer.deallocate`.

This presented itself as a heap buffer corruption on Windows when
building swift-package-manager with swift-package-manager in release
mode on Windows.