Provide capability for clients to build a node and walk the rule result info for the node and its dependencies
This is implemented by having the client optionally pass an implementation of a `RuleResultsWalker` which receives the related information for the nodes.
I preferred this method of allowing the client to "hook" in and get the info before the `build` call returns and does its clean up and any related "teardown".
That way we ensure the data is readily available for the client to inspect.
This method also allows to restrict the rule result queries only to the nodes related to the just built node, which simplifies things compared to exposing general APIs for querying any node, even if it wasn't involved in the build.

Related to rdar://67816715