Scope build debugging trace file to the build operation
Clients who cache and re-use a BuildSystem instance to serve multiple subsequent build operations are unable to get a trace file for multiple builds.

This is because the trace file pointer is scoped to and kept open for the lifetime of the BuildSystem instance, rather than being scoped to the build operation.

This change scopes the trace file writing to the build operation. This means that multiple build operations by the same BuildSystem will no longer be able to "append" to the same trace file (e.g. having multiple build-started/build-ended pairs), but this responsibility can be shifted to clients who can preserve the file between builds if desired.

rdar://47133775