Add a C API and corresponding Swift interface for the Ninja component
The current Swift Ninja library runs the `llbuild` process to output a
JSON file and then reads it in. Add a new C API to interface with the
llbuild Ninja component. Use this to create a NinjaManifest interface in
the Swift llbuild library and remove the separate Ninja library.

Some known missing functionality:
  - Variables in the top-level/sub scopes
  - Cannot read all rules (the Swift interface exposes a map for rules
    that are referenced from statements, but will miss unreferenced rules)
  - Accessors to various special variables in a command (though they can
    be manually found through the `variables` field)