[BuildSystemDelegate] Introduce `BuildSystemDelegate::chooseCommandFromMultipleProducers()`
For certain kinds of build operations, it is useful to be able to continue even after errors are occurring.
For the case of a node having multiple producers, give the option to the client delegate to choose the most appropriate command,
and allow the build to continue.

Part of rdar://80341017