[Xcode] Updated the "Install in CommandLineTools" shell script build phase to not do anything meaningful if `COMMANDLINETOOLS_FRAMEWORK_INSTALL_PATH` is unset.
`COMMANDLINETOOLS_FRAMEWORK_INSTALL_PATH` is a new build setting indicating where in the CommandLineTools package llbuild.framework should be installed.
This is configured in Common.xcconfig, allowing clients to customize this location via CommonOverrides.xcconfig.
Additionally, if set to the empty string, the "Install in CommandLineTools" build phase will effectively do nothing.

As part of this, marked the "Install in CommandLineTools" build phase so it only runs during install builds.

This is for <rdar://problem/84568181>.