build: make `libllbuild` static
Rather than emitting `libllbuild` as a dynamic library, expose it as a
static library.  This is used by a single user (`llbuildSwift`), and
thus allows for simplification of the distribution.  We no longer export
libllbuild as the clients are using llbuildSwift.  Because the module is
not imported as `@_implementationOnly`, we need to expose the include
path for `libllbuild` to clients of `llbuildSwift`.