Handle EINTR and EAGAIN when we `poll`
# Motivation
We are currently making calls to `poll` which could get interrupted or the resource could be temporarily not available. We should handle these cases.

# Modification
Wrap the calls to `poll` in a `wile` loop and try again if we get either `EINTR` or `EAGAIN`.

# Result
We are now handling signals while we are polling correctly.