Introduce the concept of "detached" external commands for the C/Swift API
These are commands that run without blocking the execution lanes or being constrained by them
(the number of execution lanes does not restrict how many 'detached' commands can run in parallel).

It is up to the client to decide whether a 'detached' command is appropriate for its purposes, but a good
use case for executing commands outside the execution lanes is when the work involves the network; in such case
it is beneficial to avoid blocking execution lanes while waiting for network queries, while also being able to initiate more
parallel network queries than the number of CPU cores.