Drop policy CMP0037 OLD setting
CMP0037 was set to avoid warnings about using reserved target names
(`test`). As of CMake 3.11, the test keyword is only reserved when the
`ctest` module or testing is enabled through CMake. The llbuild build
system requires at least 3.19, so this setting now just results in a
warning about how the old behavior is deprecated.