[FreeBSD] Fix target conditionals for FreeBSD aarch64 builds
Despite being declared in the header unistd.h, the function sbrk is not available on FreeBSD when using the aarch64 architecture because FreeBSD make the decision to drop support for this deprecated function when porting the system to ARM.

The configuration script incorrectly hard-codes the support for this function as being always available, which is not correct. We need to ensure that it is not used on FreeBSD aarch64.