llbuild: guard the complete header from duplicate inclusion
The header guards prevent re-processing the header when re-referenced.
We were not guarding the entire file. Although this matters very little
in practice in this library as the headers will not change behaviour and
inclusion into Swift will go through a module, this is a small drive-by
clean up that makes code more idiomatic.