build: rename `libllbuild` to `llbuild`, `llbuild` to `llbuild-tool`
Because `llbuild` is the executable, we cannot have another target with
the name `llbuild` which is the library target. As such, the library is
named `libllbuild`. This changes the output name of the library to
`liblibllbuild`. This was being overridden which caused some trouble for
builds which attempted to skirt the dependencies.

As the tool is generally unused, prefer to rename the tool target to
`llbuild-tool`, and use `OUTPUT_NAME` to retain the output name. This
allows us to have the `libllbuild.a` name for the new `llbuild` target.